Add wordCount property closes #16
It's a basic functionality, and should be reviewed, there's a kind of
bug already, if the user select something like ', a', wordCount will
return two words selected, but, actually there's just one word in this
string. We should remove punctuation and other characters that ins't a
word.

But, it's working.